requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pyyaml==6.0.1
python-magic==0.4.27
bandit==1.7.5
pylint==3.0.3
astroid==3.0.2
colorama==0.4.6
click==8.1.7
